Extract a reusable CucumberJava8ClassVisitor
